@@ -4,18 +4,18 @@ version: 1.0.0
 homepage: https://github.com/jamieastley/pop_pop
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  rxdart: 0.24.1
+  rxdart: ^0.27.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.2
-  pedantic: ^1.9.2
+  mockito: ^5.0.16
+  flutter_lints: ^1.0.4
 
 flutter:
