@@ -10,8 +10,8 @@ class PoppedBubblePainter extends StatelessWidget {
   final PopPopBubbleTheme themeModel;
 
   const PoppedBubblePainter({
-    Key key,
-    @required this.themeModel,
+    Key? key,
+    required this.themeModel,
   }) : super(key: key);
 
   @override
@@ -26,14 +26,14 @@ class PoppedBubblePainter extends StatelessWidget {
 class _Painter extends CustomPainter {
   final PopPopBubbleTheme themeModel;
 
-  const _Painter({@required this.themeModel});
+  const _Painter({required this.themeModel});
 
   @override
   void paint(Canvas canvas, Size size) {
-    var painterSize = (size.height * 0.85) - themeModel.strokeWidth;
+    final painterSize = (size.height * 0.85) - themeModel.strokeWidth;
 
-    var center = size.center(Offset.zero);
-    var offset = size.height * 0.05;
+    final center = size.center(Offset.zero);
+    final offset = size.height * 0.05;
 
     // ? base bubble layer
     canvas.drawCircle(
